requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
transformers
datasets
sentence-transformers
torch
torchvision
timm
Pillow
scikit-learn
joblib
pandas
numpy
evaluate
requests
